task/35: chore - refactor folder structure in API to be more consistent and update to use absolute import
