+++ apps/api/src/index.ts
@@ -1,5 +1,6 @@
+import { log } from "infra/logger";
+
 import { createServer } from "./server";
-import { log } from "./infra/logger.ts";
 
 const port = process.env.PORT || 5001;
 const server = createServer();
