--- apps/api/src/infra/database-client.ts
@@ -1,5 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-const DatabaseClient = new PrismaClient();
-
-export default DatabaseClient;
