+++ apps/api/src/infra/db-client.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from "@prisma/client";
+
+const dbClient = new PrismaClient();
+
+export default dbClient;
