+++ apps/api/src/routes/auth/handlers/index.ts
@@ -1,0 +1,2 @@
+export * from "./login.handler";
+export * from "./register.handler";
