+++ apps/api/src/routes/auth/helpers/auth.helper.ts
