+++ apps/api/src/routes/auth/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from "./auth.helper";
