+++ apps/api/src/routes/auth/index.ts
@@ -1,5 +1,2 @@
-import loginHandler from './login.handler';
-import registerHandler from './register.handler';
-import tokenMiddleware from './token.middleware';
-
-export { loginHandler, registerHandler, tokenMiddleware };
+export { authMiddleware } from "./middlewares";
+export { default } from "./router";
