+++ apps/api/src/routes/auth/repositories/index.ts
@@ -1,0 +1,1 @@
+export * as userRepository from "./user.repository";
