+++ apps/api/src/routes/auth/router.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+
+import { loginHandler, registerHandler } from "./handlers";
+
+const router = Router();
+
+router.post("/login", loginHandler);
+router.post("/register", registerHandler);
+
+export default router;
