+++ apps/api/src/routes/keywords/handlers/index.ts
@@ -1,0 +1,2 @@
+export * from "./get-keyword.handler";
+export * from "./get-keywords.handler";
