+++ apps/api/src/routes/keywords/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./handler";
+export { default } from "./router";
