+++ apps/api/src/routes/keywords/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from "./keyword.repository";
