+++ apps/api/src/routes/uploads/handlers/index.ts
@@ -1,0 +1,2 @@
+export { createNewUploadHandler } from "./create-upload.handler";
+export { getUploadsHandler } from "./get-uploads.handler";
