+++ apps/api/src/routes/uploads/helpers/index.ts
@@ -1,0 +1,2 @@
+export * as scraper from "./scraper";
+export * as resultProcessor from "./result-processor";
