+++ apps/api/src/routes/uploads/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./handler";
+export { default } from "./router";
