+++ apps/api/src/routes/uploads/middlewares/index.ts
@@ -1,0 +1,1 @@
+export * from "./upload.middleware";
