+++ apps/api/src/routes/uploads/repositories/index.ts
@@ -1,0 +1,1 @@
+export * as uploadRepository from "./upload.repository";
