+++ apps/api/src/server.ts
@@ -3,10 +3,10 @@ import express, { type Express } from "express";
 import morgan from "morgan";
 import compression from "compression";
 
-import cors, { handleError } from "./middlewares";
-import { loginHandler, registerHandler, tokenMiddleware } from "./routes/auth";
-import keywordsRouter from "./routes/keywords";
-import uploadsRouter from "./routes/uploads";
+import cors, { handleError } from "middlewares";
+import authRouter, { authMiddleware } from "routes/auth";
+import keywordsRouter from "routes/keywords";
+import uploadsRouter from "routes/uploads";
 
 export const createServer = (): Express => {
   const app = express();
@@ -20,9 +20,8 @@ export const createServer = (): Express => {
     .get("/status", (_, res) => {
       return res.json({ ok: true });
     })
-    .post("/login", loginHandler)
-    .post("/register", registerHandler)
-    .use(tokenMiddleware)
+    .use("", authRouter)
+    .use(authMiddleware)
     .use("/uploads", uploadsRouter)
     .use("/keywords", keywordsRouter)
     .use(handleError);
