+++ packages/tsconfig/base.json
@@ -7,7 +7,8 @@
     "declarationMap": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
-    "allowImportingTsExtensions": true,
+    "allowImportingTsExtensions": false,
+    "baseUrl": "./src",
     "inlineSources": false,
     "isolatedModules": true,
     "moduleResolution": "node",
